--- src/climate_learn/models/modules/__init__.py
@@ -1,2 +1,0 @@
-from .forecast import ForecastLitModule
-from .downscale import DownscaleLitModule
